+++ docs/utils/mk_quick_index.py
@@ -67,6 +67,22 @@ SOURCE_FILES = {
                              "grouping": "Standard datasets"},
     "multi_table_telephony_provider.py": {"briefDesc": "Provider for `multi-table/telephony` standard dataset",
                                "grouping": "Standard datasets"},
+    "constraint.py": {"briefDesc": "Constraint related code",
+                "grouping": "data generation constraints"},
+    "chained_relation.py": {"briefDesc": "ChainedInequality constraint related code",
+                             "grouping": "data generation constraints"},
+    "value_multiple_constraint.py": {"briefDesc": "FixedIncrement constraint related code",
+                               "grouping": "data generation constraints"},
+    "negative_values.py": {"briefDesc": "Negative constraint related code",
+                 "grouping": "data generation constraints"},
+    "positive_values.py": {"briefDesc": "Positive constraint related code",
+                                    "grouping": "data generation constraints"},
+    "literal_relation_constraint.py": {"briefDesc": "Scalar inequality constraint related code",
+                               "grouping": "data generation constraints"},
+    "literal_range_constraint.py": {"briefDesc": "ScalarRange constraint related code",
+                                        "grouping": "data generation constraints"},
+    "sql_expr.py": {"briefDesc": "SQL expression constraint related code",
+                                        "grouping": "data generation constraints"},
 
 }
 
@@ -81,6 +97,9 @@ GROUPING_INFO = {
     },
     "data distribution": {
         "heading": "Data distribution related classes, functions and types"
+    },
+    "data generation constraints": {
+        "heading": "Data generation constraints related classes, functions and types"
     }
 }
 
@@ -307,11 +326,17 @@ def main(dirToSearch, outputPath):
         writeUnderlined(outputFile, f"The ``{PACKAGE_NAME}`` package", underline="_")
         processDirectory(outputFile, Path(f"{PROJECT_PATH}"))
 
-        writeUnderlined(outputFile, f"The ``{PACKAGE_NAME}.distributions`` package", underline="_")
-        processDirectory(outputFile, Path(f"{PROJECT_PATH}/distributions"), subpackage="distributions")
+        # Add entries here for subpackages
+        writeUnderlined(outputFile, f"The ``{PACKAGE_NAME}.constraints`` package", underline="_")
+        processDirectory(outputFile, Path(f"{PROJECT_PATH}/constraints"), subpackage="constraints")
 
         writeUnderlined(outputFile, f"The ``{PACKAGE_NAME}.datasets`` package", underline="_")
         processDirectory(outputFile, Path(f"{PROJECT_PATH}/datasets"), subpackage="datasets")
+
+        writeUnderlined(outputFile, f"The ``{PACKAGE_NAME}.distributions`` package", underline="_")
+        processDirectory(outputFile, Path(f"{PROJECT_PATH}/distributions"), subpackage="distributions")
+
+
 
 if __name__ == '__main__':
     main(sys.argv[1], sys.argv[2])